refactor: 피드 디렉터리 합치기
